#297 refactor(shutdown): Put everything in a services package
